project/mini-amazon-ads-ai/mini-amazon-ads-ai/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
clickhouse-connect==0.8.5
anthropic==0.39.0
httpx==0.27.2
email-validator==2.2.0